Step 5 & Step 6
Step 5 -> Calling Add with a negative number will throw an exception “negatives not allowed” - and the negative that was passed.

If there are multiple negatives, show all of them in the exception message.

Step 6 ->
Numbers bigger than 1000 should be ignored, so adding 2 + 1001 = 2

I added 2 steps at once, since I forgot to commit earlier step 5 and don't want to go back, commit and then add functionality to Step 5 to be Step 6 and commit.